+++ crates/solana-cli/src/command/shreds/list.rs
@@ -68,7 +68,7 @@ struct SeatRow {
 impl ListCommand {
     pub async fn try_into_execute(self, dz_ledger_url: Option<String>) -> Result<()> {
         let moniker_env = self.connection_options.moniker_env();
-        let connection = self.connection_options.into_shred_subscription_connection();
+        let connection = self.connection_options.into_solana_connection();
 
         let discriminator_bytes =
             borsh::to_vec(&state::CLIENT_SEAT_DISCRIMINATOR).expect("discriminator serialization");