+++ crates/solana-cli/src/command/shreds/publisher_rewards/init.rs
@@ -36,7 +36,7 @@ impl InitCommand {
             .solana_payer_options
             .connection_options
             .clone()
-            .into_shred_subscription_connection();
+            .into_solana_connection();
         let wallet = Wallet::try_new(self.solana_payer_options, Some(dz_connection))?;
         let wallet_key = wallet.pubkey();