+++ crates/solana-cli/src/command/shreds/validator_client_rewards/show.rs
@@ -135,7 +135,7 @@ impl ShowCommand {
             .solana_payer_options
             .connection_options
             .clone()
-            .into_shred_subscription_connection();
+            .into_solana_connection();
         let wallet = Wallet::try_new(self.solana_payer_options, Some(dz_connection))?;
 
         let vcr_key = find_validator_client_rewards_address(self.client_id).0;